aig/src/requirements.txt
setuptools==80.9.0
# OpenVINO and Intel optimizations
openvino_genai==2025.4.1.0
numpy==2.2.6
pillow==11.2.1
diffusers==0.33.1
flask==2.2.5
flask-restx==1.3.0
chromadb==1.3.7
sentence-transformers==5.2.0